Introduce base test classes for LMDB and InMemory Stores
Separate IStoreTests from StorageActionsTests
