+++ rspace/src/test/scala/coop/rchain/rspace/StorageTestsBase.scala
@@ -1,7 +1,12 @@
 package coop.rchain.rspace
 
+import java.nio.file.{Files, Path}
+
 import com.typesafe.scalalogging.Logger
-import org.scalatest.{FlatSpec, Matchers, OptionValues, Outcome}
+import coop.rchain.rspace.examples.StringExamples._
+import coop.rchain.rspace.examples.StringExamples.implicits._
+import coop.rchain.rspace.test._
+import org.scalatest._
 
 trait StorageTestsBase[C, P, A, K] extends FlatSpec with Matchers with OptionValues {
 
@@ -17,4 +22,38 @@ trait StorageTestsBase[C, P, A, K] extends FlatSpec with Matchers with OptionVal
   /** A fixture for creating and running a test with a fresh instance of the test store.
     */
   def withTestStore(f: T => Unit): Unit
+}
+
+class InMemoryStoreTestsBase extends StorageTestsBase[String, Pattern, String, StringsCaptor] {
+
+  override def withTestStore(f: T => Unit): Unit = {
+    val testStore = InMemoryStore.create[String, Pattern, String, StringsCaptor]
+    testStore.clear()
+    try {
+      f(testStore)
+    } finally {
+      testStore.close()
+    }
+  }
+}
+
+class LMDBStoreTestsBase
+    extends StorageTestsBase[String, Pattern, String, StringsCaptor]
+    with BeforeAndAfterAll {
+
+  val dbDir: Path   = Files.createTempDirectory("rchain-storage-test-")
+  val mapSize: Long = 1024L * 1024L * 1024L
+
+  override def withTestStore(f: T => Unit): Unit = {
+    val testStore = LMDBStore.create[String, Pattern, String, StringsCaptor](dbDir, mapSize)
+    testStore.clear()
+    try {
+      f(testStore)
+    } finally {
+      testStore.close()
+    }
+  }
+
+  override def afterAll(): Unit =
+    recursivelyDeletePath(dbDir)
 }
